feat: added --refresh option to spar extract

+++ data_prep/db_env_utils/object_store.py
@@ -4,7 +4,9 @@ Wrapper to object storage functionality.
 
 from __future__ import annotations
 
+import hashlib
 import logging
+import time
 from typing import TYPE_CHECKING
 
 import boto3
@@ -122,8 +124,9 @@ class OStore:
         :rtype: bool
         """
         try:
+            LOGGER.debug("bucket: %s", self.conn_params.bucket)
             self.s3_client.get_object(
-                Bucket=self._bucket,
+                Bucket=self.conn_params.bucket,
                 Key=object_name,
             )
             return True  # noqa: TRY300
@@ -208,6 +211,13 @@ class OStore:
             )
         return versions
 
+    def calculate_sha256(self, file_path: pathlib.Path) -> str:
+        sha256_hash = hashlib.sha256()
+        with file_path.open("rb") as f:
+            for chunk in iter(lambda: f.read(4096), b""):
+                sha256_hash.update(chunk)
+        return sha256_hash.hexdigest()
+
     def put_data_files(
         self,
         tables: list[str],
@@ -238,14 +248,35 @@ class OStore:
                 table,
                 db_type,
             )
+            LOGGER.debug("local file: %s", local_data_file)
+            LOGGER.debug("remote file: %s", remote_data_file)
 
             # keeping it simple for now, if local exists re-use it
-            if local_data_file.exists():
+            if local_data_file.exists() and self.object_exists(
+                object_name=str(remote_data_file),
+            ):
+                LOGGER.debug(
+                    "delete pre-existing remote file: %s", remote_data_file
+                )
                 self.delete_data_file(remote_data_file)
                 # pull the files from object store.
+            LOGGER.debug(
+                "uploading file: %s to: %s in the bucket %s",
+                local_data_file,
+                remote_data_file,
+                self.conn_params.bucket,
+            )
+
+            # when the file gets deleted for some reason was generating a checksum
+            # error when trying to re-upload it... Only way to fix was to
+            # calculate the checksum locally and send it along with the upload.
+            checksum = self.calculate_sha256(file_path=local_data_file)
+            LOGGER.debug("checksum: %s", checksum)
+
             response = self.s3_client.upload_file(
                 str(local_data_file),
                 self.conn_params.bucket,
                 str(remote_data_file),
+                ExtraArgs={"ChecksumSHA256": checksum},
             )
             LOGGER.debug("response from object store upload: %s", response)
